requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
arabic-reshaper==2.1.1
asgiref==3.3.1
asn1crypto==1.5.1
attrs==20.3.0
autopep8==1.5.4
boto3==1.12.6
botocore==1.15.49
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==8.1.3
cryptography==3.3.1
cssselect2==0.7.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.10
django-absoluteuri==1.3.0
django-allauth==0.44.0
django-anymail==8.1
django-apscheduler==0.5.1
django-cleanup==4.0.0
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-qr-code==2.0.1
django-rest-auth==0.9.5
django-simple-history==2.12.0
django-storages==1.11
django-widget-tweaks==1.4.8
djangoconvertvdoctopdf==1.0.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
docutils==0.15.2
docxtpl==0.11.2
drf-extensions==0.6.0
et-xmlfile==1.0.1
filetype==1.0.7
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
isodate==0.6.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.24.0
oauthlib==3.1.0
openpyxl==3.0.6
oscrypto==1.3.0
Pillow==9.3.0
platformdirs==2.6.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.16.0
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==1.7.1
PyPDF3==1.0.6
PyQRCode==1.2.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.6
python-docx==0.8.10
python3-openid==3.2.0
pytz==2020.4
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
s3transfer==0.3.3
segno==1.3.1
six==1.15.0
sqlparse==0.4.1
svglib==1.4.1
tinycss2==1.2.1
toml==0.10.2
tqdm==4.64.1
tzlocal==2.1
Unipath==1.1
uritools==4.0.0
urllib3==1.25.11
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.8
xlwt==1.3.0
xmltodict==0.13.0
zeep==4.2.1
